feat: add fetch error handling
