+++ src/App.css
@@ -5,16 +5,20 @@
   min-height: 100vh;
 }
 
-.main {
+.main, .error {
   flex-grow: 1;
   padding: 20px;
+  color: red;
+  font-size: 2rem;
 }
 
-.loader-container {
+.loader-container, .no-results {
   position: fixed;
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
+  color: white;
+  font-size: 3rem;
 }
 
 .loader {
@@ -23,24 +27,38 @@
   display: grid;
   border-radius: 50%;
   background:
-    linear-gradient(0deg ,rgba(236, 232, 232, 0.5) 30%,#9b040400 0 70%,rgb(128, 136, 129) 0) 50%/8% 100%,
-    linear-gradient(90deg,rgba(129, 124, 124, 0.25) 30%,#b9090900 0 70%,rgba(117, 108, 108, 0.75) 0) 50%/100% 8%;
+    linear-gradient(
+        0deg,
+        rgba(236, 232, 232, 0.5) 30%,
+        #9b040400 0 70%,
+        rgb(128, 136, 129) 0
+      )
+      50%/8% 100%,
+    linear-gradient(
+        90deg,
+        rgba(129, 124, 124, 0.25) 30%,
+        #b9090900 0 70%,
+        rgba(117, 108, 108, 0.75) 0
+      )
+      50%/100% 8%;
   background-repeat: no-repeat;
   animation: l23 1s infinite steps(12);
 }
 .loader::before,
 .loader::after {
-   content: "";
-   grid-area: 1/1;
-   border-radius: 50%;
-   background: inherit;
-   opacity: 0.915;
-   transform: rotate(30deg);
+  content: '';
+  grid-area: 1/1;
+  border-radius: 50%;
+  background: inherit;
+  opacity: 0.915;
+  transform: rotate(30deg);
 }
 .loader::after {
-   opacity: 0.83;
-   transform: rotate(60deg);
+  opacity: 0.83;
+  transform: rotate(60deg);
 }
 @keyframes l23 {
-  100% {transform: rotate(1turn)}
+  100% {
+    transform: rotate(1turn);
+  }
 }
