+++ src/components/footer/styles.css
@@ -1,7 +1,6 @@
 .footer {
   width: 100%;
   min-height: 50px;
-
 }
 
 .footer a {
