+++ src/components/header/Header.tsx
@@ -5,7 +5,7 @@ export class Header extends Component {
   render() {
     return (
       <header className="header">
-        <h1 className="title">People from Star Wars</h1>
+        <h1 className="title">Star Wars Сharacters</h1>
       </header>
     );
   }
