+++ src/components/main/searchForm/styles.css
@@ -1,0 +1,17 @@
+.search-form {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 10px;
+    margin: 20px;
+  }
+  
+  .search-input {
+    width: 500px;
+    height: 30px;
+  }
+  
+  .search-button {
+    width: 100px;
+    height: 30px;
+  }
